+++ oap-server/server-configuration/configuration-api/src/main/java/org/apache/skywalking/oap/server/configuration/api/ConfigWatcherRegister.java
@@ -35,8 +35,9 @@ import org.slf4j.LoggerFactory;
 public abstract class ConfigWatcherRegister implements DynamicConfigurationService {
     private static final Logger LOGGER = LoggerFactory.getLogger(ConfigWatcherRegister.class);
     public static final String LINE_SEPARATOR = System.getProperty("line.separator", "\n");
+    private Register singleConfigChangeWatcherRegister = new Register();
     @Getter
-    private Register register = new Register();
+    private Register groupConfigChangeWatcherRegister = new Register();
     private volatile boolean isStarted = false;
     private final long syncPeriod;
 
@@ -55,16 +56,28 @@ public abstract class ConfigWatcherRegister implements DynamicConfigurationServi
         }
 
         WatcherHolder holder = new WatcherHolder(watcher);
-        if (register.containsKey(holder.getKey())) {
+        if (singleConfigChangeWatcherRegister.containsKey(
+            holder.getKey()) || groupConfigChangeWatcherRegister.containsKey(holder.getKey())) {
             throw new IllegalStateException("Duplicate register, watcher=" + watcher);
         }
-        register.put(holder.getKey(), holder);
+
+        switch (holder.getWatcher().getWatchType()) {
+            case SINGLE:
+                singleConfigChangeWatcherRegister.put(holder.getKey(), holder);
+                break;
+            case GROUP:
+                groupConfigChangeWatcherRegister.put(holder.getKey(), holder);
+                break;
+            default:
+                throw new IllegalArgumentException(
+                    "Unexpected watch type of ConfigChangeWatcher " + watcher.toString());
+        }
     }
 
     public void start() {
         isStarted = true;
 
-        LOGGER.info("Current configurations after the bootstrap sync." + LINE_SEPARATOR + register.toString());
+        LOGGER.info("Current configurations after the bootstrap sync." + LINE_SEPARATOR + singleConfigChangeWatcherRegister.toString());
 
         Executors.newSingleThreadScheduledExecutor()
                  .scheduleAtFixedRate(
@@ -75,13 +88,18 @@ public abstract class ConfigWatcherRegister implements DynamicConfigurationServi
     }
 
     void configSync() {
-        Optional<ConfigTable> configTable = readConfig(register.keys());
+        singleConfigsSync();
+        groupConfigsSync();
+    }
+
+    private void singleConfigsSync() {
+        Optional<ConfigTable> configTable = readConfig(singleConfigChangeWatcherRegister.keys());
 
         // Config table would be null if no change detected from the implementation.
         configTable.ifPresent(config -> {
             config.getItems().forEach(item -> {
                 String itemName = item.getName();
-                WatcherHolder holder = register.get(itemName);
+                WatcherHolder holder = singleConfigChangeWatcherRegister.get(itemName);
                 if (holder == null) {
                     LOGGER.warn(
                         "Config {} from configuration center, doesn't match any WatchType.SINGLE watcher, ignore.",
@@ -110,71 +128,86 @@ public abstract class ConfigWatcherRegister implements DynamicConfigurationServi
                     }
                 }
             });
-
-            LOGGER.trace("Current configurations after the sync." + LINE_SEPARATOR + register.toString());
+            if (LOGGER.isTraceEnabled()) {
+                LOGGER.trace(
+                    "Current configurations after the sync." + LINE_SEPARATOR + singleConfigChangeWatcherRegister.toString());
+            }
         });
-
-        configTable.ifPresent(this::groupConfigsSync);
     }
 
-    private void groupConfigsSync(ConfigTable config) {
-        config.getGroupItems().forEach(groupConfigItems -> {
-            String groupConfigItemName = groupConfigItems.getName();
-            WatcherHolder holder = register.get(groupConfigItemName);
+    private void groupConfigsSync() {
+        Optional<GroupConfigTable> groupConfigTable = readGroupConfig(groupConfigChangeWatcherRegister.keys());
+        // Config table would be null if no change detected from the implementation.
+        groupConfigTable.ifPresent(config -> {
+            config.getGroupItems().forEach(groupConfigItems -> {
+                String groupConfigItemName = groupConfigItems.getName();
+                WatcherHolder holder = groupConfigChangeWatcherRegister.get(groupConfigItemName);
 
-            if (holder == null || holder.getWatcher().watchType != ConfigChangeWatcher.WatchType.GROUP) {
-                LOGGER.warn(
-                    "Config {} from configuration center, doesn't match any WatchType.GROUP watcher, ignore.",
-                    groupConfigItemName
-                );
-                return;
-            }
+                if (holder == null) {
+                    LOGGER.warn(
+                        "Config {} from configuration center, doesn't match any WatchType.GROUP watcher, ignore.",
+                        groupConfigItemName
+                    );
+                    return;
+                }
 
-            GroupConfigChangeWatcher watcher = (GroupConfigChangeWatcher) holder.getWatcher();
-            Map<String, ConfigTable.ConfigItem> groupItems = groupConfigItems.getItems();
-            groupItems.forEach((groupItemName, groupItem) -> {
-                String newItemValue = groupItem.getValue();
-                if (newItemValue == null) {
-                    if (watcher.groupItems().get(groupItemName) != null) {
-                        // Notify watcher, the new value is null with delete event type.
-                        watcher.notify(
-                            new ConfigChangeWatcher.ConfigChangeEvent(
-                                groupItemName,
+                GroupConfigChangeWatcher watcher = (GroupConfigChangeWatcher) holder.getWatcher();
+                Map<String, ConfigTable.ConfigItem> groupItems = groupConfigItems.getItems();
+                Map<String, ConfigChangeWatcher.ConfigChangeEvent> changedGroupItems = new HashMap<>();
+                Map<String, String> currentGroupItems = Optional.ofNullable(watcher.groupItems())
+                                                                .orElse(new HashMap<>());
+
+                groupItems.forEach((groupItemName, groupItem) -> {
+                    String newItemValue = groupItem.getValue();
+                    if (newItemValue == null) {
+                        if (currentGroupItems.get(groupItemName) != null) {
+                            // Notify watcher, the new value is null with delete event type.
+                            changedGroupItems.put(groupItemName, new ConfigChangeWatcher.ConfigChangeEvent(
                                 null,
                                 ConfigChangeWatcher.EventType.DELETE
                             ));
-                    } else {
-                        // Don't need to notify, stay in null.
-                    }
-                } else { //add and modify
-                    if (!newItemValue.equals(watcher.groupItems().get(groupItemName))) {
-                        watcher.notify(new ConfigChangeWatcher.ConfigChangeEvent(
-                            groupItemName,
-                            newItemValue,
-                            ConfigChangeWatcher.EventType.MODIFY
-                        ));
-                    } else {
-                        // Don't need to notify, stay in the same config value.
-                    }
-                }
-            });
 
-            watcher.groupItems().forEach((oldGroupItemName, oldGroupItemValue) -> {
-                //delete item
-                if (null == groupItems.get(oldGroupItemName)) {
-                    // Notify watcher, the item is deleted with delete event type.
-                    watcher.notify(
-                        new ConfigChangeWatcher.ConfigChangeEvent(oldGroupItemName,
-                                                                  null, ConfigChangeWatcher.EventType.DELETE
+                        } else {
+                            // Don't need to notify, stay in null.
+                        }
+                    } else { //add and modify
+                        if (!newItemValue.equals(currentGroupItems.get(groupItemName))) {
+                            changedGroupItems.put(groupItemName, new ConfigChangeWatcher.ConfigChangeEvent(
+                                newItemValue,
+                                ConfigChangeWatcher.EventType.MODIFY
+                            ));
+
+                        } else {
+                            // Don't need to notify, stay in the same config value.
+                        }
+                    }
+                });
+
+                currentGroupItems.forEach((oldGroupItemName, oldGroupItemValue) -> {
+                    //delete item
+                    if (null == groupItems.get(oldGroupItemName)) {
+                        // Notify watcher, the item is deleted with delete event type.
+                        changedGroupItems.put(oldGroupItemName, new ConfigChangeWatcher.ConfigChangeEvent(
+                            null,
+                            ConfigChangeWatcher.EventType.DELETE
                         ));
+                    }
+                });
+
+                if (changedGroupItems.size() > 0) {
+                    watcher.notifyGroup(changedGroupItems);
                 }
             });
+            if (LOGGER.isTraceEnabled()) {
+                LOGGER.trace(
+                    "Current configurations after the sync." + LINE_SEPARATOR + groupConfigChangeWatcherRegister.toString());
+            }
         });
-
-        LOGGER.trace("Current configurations after the sync." + LINE_SEPARATOR + register.toString());
     }
 
     public abstract Optional<ConfigTable> readConfig(Set<String> keys);
+
+    public abstract Optional<GroupConfigTable> readGroupConfig(Set<String> keys);
 
     public class Register {
         private Map<String, WatcherHolder> register = new HashMap<>();
@@ -208,14 +241,13 @@ public abstract class ConfigWatcherRegister implements DynamicConfigurationServi
                                         .append(watcher.getModule())
                                         .append("    provider:")
                                         .append(watcher.getProvider().name());
-                if (watcher.watchType == ConfigChangeWatcher.WatchType.GROUP) {
+                if (watcher.watchType.equals(ConfigChangeWatcher.WatchType.GROUP)) {
                     GroupConfigChangeWatcher groupWatcher = (GroupConfigChangeWatcher) watcher;
                     registerTableDescription.append("    groupItems(current):")
                                             .append(groupWatcher.groupItems());
                 } else {
                     registerTableDescription.append("    value(current):")
                                             .append(watcher.value());
-
                 }
                 registerTableDescription.append(LINE_SEPARATOR);
             });